requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain_openai    
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-ollama
streamlit
langchain_groq